Add oh-my-zsh-activate script
Script: oh-my-zsh-activate
Description: This script will look for Oh My Zsh directory, and install it if not found, or load it up and switch to ZSH if you are using other terminal colorizing tool on non-zsh terminal. Check https://ohmyz.sh for more information about the Oh My Zsh for colorizing, and framework management for your Zsh configuration.